Make denote-sequence-find-{next,previous}-sibling reuse found siblings
This builds on commit ea4a9d3 by Ashton Wiersdorf, which was
contributed as part of pull request 11: <https://github.com/protesilaos/denote-sequence/pull/11>.

The idea of that commit is to keep searching when we encounter a gap
in the sequence, like in "1a, 1b, 1d". Given that the list of siblings
should be the same as we go through them, we can keep using the
list of files we originally got.