Implement command to view hierarchy.el tree (BUT I WILL CHANGE IT)
I am committing this to have a record of it. But I am not happy with
what hierarchy.el provides. The tree widget it has lacks many features
I want to see in such an interface, plus it works with assumptions
that are not useful for my purposes. I will either remove this command
altogether, or implement what I need myself.